requirements.txt
# Core Django Dependencies
Django>=4.2.0
djangorestframework>=3.14.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Image Processing
Pillow>=9.5.0

# Environment Variables
python-dotenv>=1.0.0

# AI and Machine Learning Dependencies
google-generativeai>=0.3.0
scikit-learn>=1.2.0
nltk>=3.8.1
transformers>=4.30.0
torch>=2.0.0

# Frontend Dependencies
requests>=2.31.0
urllib3>=2.0.7

# Development Tools
django-debug-toolbar>=4.2.0
black>=23.11.0  # Code formatting
flake8>=6.1.0   # Code linting
pytest>=7.4.3    # Testing
pytest-django>=4.7.0

# Security
django-cors-headers>=4.3.1
python-jose>=3.3.0  # JWT tokens
bcrypt>=4.0.1       # Password hashing